backend/requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx

# For mock data generation
Faker
PyJWT
